requirements.txt
torch==2.3.0
torchvision==0.18.0
torchdiffeq==0.2.3
deepxde==1.11.1
diffusers==0.27.2
transformers==4.40.0
langchain==0.1.20
langgraph==0.0.55
langchain-openai==0.1.6
filterpy==1.4.5
numpy==1.26.4
scipy==1.13.0
matplotlib==3.8.4
plotly==5.22.0
wandb==0.16.6
pyyaml==6.0.1
python-dotenv==1.0.1
tqdm==4.66.4
pytest==8.2.0
black==24.4.2
flake8==7.0.0
jupyter==1.0.0
pandas==2.2.2
netCDF4==1.6.5
h5py==3.11.0
scikit-learn==1.4.2
onnx==1.16.0
onnxruntime==1.18.0
